Teapot: OpenGL version
- 3.3 to 4.6
- GLAD update

+++ Teapot/src/Renderer/FrameBuffer.h
@@ -8,7 +8,7 @@ namespace Teapot
 	class FrameBuffer
 	{
 	public:
-		FrameBuffer(float width, float height);
+		FrameBuffer(int width, int height);
 		~FrameBuffer();
 		unsigned int GetFrameTexture() const;
 		void RescaleFrameBuffer(int width, int height) const;
